+++ packages/create-spree-app/src/types.ts
@@ -7,4 +7,5 @@ export interface ScaffoldOptions {
   start: boolean
   packageManager: PackageManager
   port: number
+  dbPort: number
 }